+++ server/db/update_scripts/16_update_v6.1.0_v6.2.0.sql
@@ -1,0 +1,8 @@
+-- set DB version
+UPDATE version
+SET version = 'v6.2.0';
+
+ -- add keyword and param columns to table user_activity
+ CALL add_table_column('user_activity', 'keyword', "VARCHAR(100) DEFAULT NULL");
+ CALL add_table_column('user_activity', 'params', "VARCHAR(1000) DEFAULT NULL");
+ CALL add_table_column('user_activity', 'mobile', "BOOLEAN DEFAULT NULL");