+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### New features
  - make function `calc_dynamic_values` protected in `StyleModel`
 
+###  Bugfix
+ - save external data on update properly add new columns if they do not exists
+
 # v6.1.6
 ###  Bugfix
  - adding input fields as children in style `validate` are saved in the DB as user input data