+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ###  Bugfix
  - save external data on update properly add new columns if they do not exists
+ - fix the size of the `markdown` fields in a style in the CMS
 
 # v6.1.6
 ###  Bugfix