+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v6.1.7 (Unpublished)
 ### New features
  - make function `calc_dynamic_values` protected in `StyleModel`
+ - add class `selfHelp-locale-language_code` example: `selfHelp-locale-de-CH` in the style
 
 ###  Bugfix
  - save external data on update properly add new columns if they do not exists