+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# v6.1.7 (Unpublished)
+# v6.2.0 (Unpublished)
 ### New features
  - make function `calc_dynamic_values` protected in `StyleModel`
  - add class `selfHelp-locale-language_code` example: `selfHelp-locale-de-CH` in the style
  - in `admin/data` hide all columns starting with "_"
+ - track activity for admin pages and show notifications when multiple people work on the same page
 
 ###  Bugfix
  - save external data on update properly add new columns if they do not exists